WEBAPP-613: implemented search for categories
